Complete removal of premature optimizations
- Replace complex pg_catalog_helpers.rs (~3000 lines) with simple version (~180 lines)
- Remove all 'enhanced', 'phase-based', and cache-dependent functions
- Eliminate 43 compile errors from removed cache dependencies
- Reduce from 137 to 104 tests (removed 33 over-engineered features)
- Keep only essential functions: get_or_allocate_type_oid, initialize_system_catalog, basic registration
- Simplify function names: register_table vs register_enhanced_table, register_database vs register_user_database
- Remove 12+ different cache types and associated metadata structs
- All tests pass - same functionality without complexity

This addresses the user's feedback that premature optimizations added
complexity without benefit since catalog loads in 2-3 seconds anyway.
